filesync/requirements.txt
pandas==2.2.1

dataclasses~=0.6

numpy==1.26.4
# DON'T UPGRADE SCHEDULE, it seems to break the plugin
schedule==1.1.0

#schedule==1.2.1

pip==24.0

datetime==5.4

psycopg2==2.9.9

sqlalchemy==2.0.27

python-dotenv~=1.0.0

setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability

pymongo~=4.8.0

django-scheduler~=0.10.1